support triggering jenkins jobs
Add support for trigggering CI (Jenkins) Jobs for PRs in the auoupdater
queue.
CI jobs are triggered when a PR is placed first in the queue and it's
check status is pending and when the branch of a PR changes.

The URL of the Jenkins CI server and the jobs that should be run are configured
in the config.toml file. The jobs are triggered by sending HTTP-Requests
to the URL of the job. It is supported to pass parameters to jobs.
Triggering jobs is retried automatically on errors.

The webinterface displays the CI configuration and configuration
settings are now shown in bold.

The main function has been refactored to:
 - start listening for github web hook events before syncing the
   autoupdater, to ensure events are received and processed that
   happened while a PR is added to the queue,
 - call autoupdater.Stop on shutdown