autoupdater: use different logfield than operation for retryer
Do not use the event logfield to assign a name for the operation that
the retryer runs. It can already be in reuse, resulting in 2 event
fields. Use a new logfield called "operation" instead.

Also fix an issue where logfields got appended to a passed logfields
slice in updatePRWithBase.
This can cause that the wrong logfields are displayed in some messages.