jenkins: interpret 302,303 and all 2xx response codes as success
The jenkins server responds with 302 when a build with the same
parameter is already scheduled.
According to https://plugins.jenkins.io/build-token-root/ it can also
return 303 for the same scenario.

Be more lenient and interpret 302,303 and all 2xx as successful
responses.
If an unexpected status code is returned a debug message is logged.