jenkins: use buildWithParameters endpoint
When triggering a job with parameters by posting to the build endpoint,
the Location header in the response is set to the Job URL.
When posting to the buildWithParameters endpoint, the location header
contains the URL to the queued item.
The URL to the queued item will be needed in a follow-up, refactor the
jenkins client accordingly.

The Client is also changed to not encode the parameters as JSON, this is
unnecessary and the buildWithParameters endpoint does not support it.

The client now also adds the /build or /buildWithParameters prefix to
the path. It must not specified in the config anymore.