jenkins: return queue item url from Build
Client.Build now returns the URL to the queued item when triggering a
build is successful.
The jenkins server sends back the url in the Location header after
posting to the build/buildWithParameters endpoint.

During tests Jenkins responded with the Job URL instead when the
endpoint was wrongly used.
Client.Build verifies that it is a URL with the path prefix /queue/item
and otherwise returns an error.

Because a build might have been scheduled when Jenkins returns a 2xx or
3xx code, no Retryable error is returned in this case.