autoupdater: ignore status of old CI builds
When CI builds are canceled because a newer build is scheduled, a failed
github status might be reported for the aborted run.
This causes that a PR in the queue is unwantedly suspended.
To prevent it, the autoupdater now ignores build status of CI build that
are known and for that a build with a higher ID has been started.

This is done by storing in the PR struct a map of the builds that have been
triggered last. If a status is received, it is checked if is a known build
and if the build id is the same or higher then the one of the started
build. If this is the case the PR is suspended. If the build id is lower
the status is ignored.

If the status for an unknown build is received or an error happens while
evaluating the status the PR is suspended and an error is logged.
Suspending it is the better option then eventually letting it block the
queue.

This also solves the issue that the commit "ci: prevent resume->suspend
loop of a PR" (8def945) addresses but more reliably without the risk to
not triggering CI jobs when it is needed.