activequeue: process PRs in fifo order
Instead of adding PRs always to the back of the active queue, they are
now inserted in order.
PRs got a new Priority field that always has the value 0. It will be
modifiable in the webinterface in a follow-up.

When inserting a new PR the order is determined by comparing it's
priority, EnqueuedAt timestamp and as fallback it's PR Number.
PRs with a higher priority, an older EnqueuedAt date and a smaller PR
number are ordered before others.
The position of the first PR in the queue is always fixed. For it CI
jobs are running and changing it's position is costly.

Because all PRs have the same priority as of now, it means PRs in the
mergequeue are processed in FIFO order.