activequeue: deprioritize handling PRs with higher suspend count
If a PR has been suspended multiple times it is likely that CI jobs for
it are failing because of an issue in the branch.
Process PRs with a higher SuspendCount but the same priority after ones
with a lower SuspendCount.
This reduces how often a PR becomes #1 in the queue when CI jobs always
fail for it and the base branch is often updated causing it to be added
again to the active queue.