webinterface: support pausing and resuming a merge-queue
Processing webhook event, checking the status and triggering CI jobs for
the first PR can now be paused and resumed in the webinterface.

Pausing the queue is handy when for example nothing should be merged
because an issue needs to be fixed in the main branch first or because
CI checks constantly fail.

When the queue is paused the queue.updatePR() method does nothing and
returns early when called. PRs can still be enqueued and dequeued.

When updating is resumed, the setStateUnchangedSince timestamp of all
PRs in the active is reset, to prevent that PRs are wrongly detected as
stale because processing was paused.
Afterwards updatePR is run for the first PR in the queue.