queue: prevent starving of PRs with suspendCount >0
If a PR has been suspended it is ordered after PRs with the same
priority in the active queue.
If there is always another PR with the same priority in the queue, PRs
with a SuspendCount != 0 are never merged.

To prevent it, only considers the SuspendCount of the PR when it has
been less than 4h in the active queue.
This still processes PRs with always failing CI check less often but
prevents that they starve in the queue when the CI checks failed one
time because of a temporary issue of a flaky test.