autoupdater: set github status state when PR is ready to be merged
When a Pullrequest is #1 in the queue, all it's required CI checks are
successful and the required number of approvals is reached a successful Github
Status state is now submitted for the head commit.
The context of the status is "directorius".
When a PR is suspended or removed from the queue and it has not been
merged the Status is set to pending.
On startup the status is set to pending for for all head commits of PRs
that are open and have a successful status state. (This resets the
status on startup.)

Github can be configured to only automerge a PR when the directorius
state is successful. This results in only merging PRs that are #1 in the
mergequeue.