autoupdater: submit successful github status on pending CI status
A successful github status is only submitted when the overall CIStatus
for the PR is successful. When Github is configured to require the
successful directorius status as merge requirement, the condition can
never be satisfied. The overall CI status is pending without the
directorius state and it is not submitted because the status is not
successful.

Submit the successful CI status for the #1 PR in the queue when it's
status is pending or successful.