queue: move evaluation logic in processPR() to separate function
The processPR() function is split to make it easier to understand and
enable testing smaller parts of it separately.

The logic that evaluates the state of a pull request is moved to the new
evalPRAction method. It returns a list of Actions that are then
performed by the processPR() function.
evalPRAction also updates a PR with it's base, split checking if an
update is necessary and triggering the update would result in more
complexity and github API calls.