githubclt: distinguish not started and running CI jobs
githubclt.ReadyForMerge() now distinguishes CI jobs that have not been
started and CI jobs that are running oer queued.
The overall CI status CIStatusExpected if a job has is not running or
queued. If all jobs are running CIStatusPending is returned.

A CI job has the status CIStatusExpected if it is a required github check and a status hasn't been reported for it or when it is commit status with state CIStatusExpected.

This will allow to only trigger CI job runs for jobs that aren't
running.