githubclt: order pull-requests by number per page in ListPRs
The graphql query used in ListPRs does not support a sort criterion that
returns a stable result for our usecase.
When using update-asc, the PR order differs often across directorius
restarts because the first PR will be updated with it's base branch or
by triggering CI jobs.
To have a more stable merge-queue order across restarts, order the PRs
per page by the pull request number.
This only works reliable when github only returns a single page for the
result. This should be most often the case because a page contains up to
100 items.

The update-asc sort criterion is kept, because it makes it more likely
that a result in a stable is returned when it is paginated.